Move the /bags API into a separate file
