+++ archive/archive_api/src/apis/__init__.py
@@ -1,0 +1,5 @@
+# -*- encoding: utf-8
+
+from .bags import api as bags_api
+
+__all__ = ["bags_api"]
